requirements.txt
torch
torchvision
Pillow
opencv-python
numpy
scikit-learn
tqdm
matplotlib
seaborn